refactor: Generator 생성자 주입 오버로딩
