+++ src/main/java/ladder/domain/LadderGame.java
@@ -5,17 +5,29 @@ import java.util.List;
 
 public class LadderGame {
 
+    private final RandomGenerator<Boolean> randomGenerator;
     private final Players players;
     private Lines lines;
 
     public LadderGame(final Players players, final int height) {
         this.players = players;
+        this.randomGenerator = new RandomBooleanGenerator();
+        initializeLines(height);
+    }
+
+    public LadderGame(final RandomGenerator<Boolean> randomGenerator, final Players players, final int height) {
+        this.players = players;
+        this.randomGenerator = randomGenerator;
         initializeLines(height);
     }
 
     private void initializeLines(final int height) {
         int width = players.size() - 1;
         this.lines = new Lines(height, width);
+    }
+
+    public void goLine() {
+
     }
 
     public List<Player> toUnmodifiablePlayers() {
