+++ src/main/java/ladder/domain/Lines.java
@@ -6,10 +6,16 @@ import java.util.List;
 
 public class Lines {
 
-    private static final RandomGenerator<Boolean> randomPointGenerator = new RandomBooleanGenerator();
+    private final RandomGenerator<Boolean> randomPointGenerator;
     private final List<Line> lines = new ArrayList<>();
 
     public Lines(final int height, final int width) {
+        this.randomPointGenerator = new RandomBooleanGenerator();
+        initialize(height, width);
+    }
+
+    public Lines(final RandomGenerator<Boolean> randomGenerator, final int height, final int width) {
+        this.randomPointGenerator = randomGenerator;
         initialize(height, width);
     }
